check null

+++ src/main/java/overcloud/blog/usecase/test/update_test/UpdateTestServiceImpl.java
@@ -127,16 +127,34 @@ public class UpdateTestServiceImpl implements UpdateTestService {
         }
 
         testRepo.updateTest(testId, request.getTitle(), request.getDescription());
-        testQuestionRepo.deleteAllById(questionsToDelete);
-        choiceAnswerRepo.deleteAll(answersToDelete);
-        practiceOpenAnswerRepo.deleteAllByQuestionId(questionsToDelete);
-        practiceChoiceRepo.deleteAll(answersToDelete);
-        questionRepo.deleteAll(questionsToDelete);
-        questionRepo.saveAll(questionsToInsert);
-        testQuestionRepo.saveAll(testQuestions(questionsToInsert, testId));
-        choiceAnswerRepo.saveAll(answersToInsert);
-        questionRepo.updateAll(questionsToUpdate);
-        choiceAnswerRepo.updateAll(answersToUpdate);    
+
+        if (questionsToDelete != null && questionsToDelete.size() > 0) {
+            testQuestionRepo.deleteAllById(questionsToDelete);
+            practiceOpenAnswerRepo.deleteAllByQuestionId(questionsToDelete);
+            questionRepo.deleteAll(questionsToDelete);
+        }
+
+        if (answersToDelete != null && answersToDelete.size() > 0) {
+            choiceAnswerRepo.deleteAll(answersToDelete);
+            practiceChoiceRepo.deleteAll(answersToDelete);
+        }
+
+        if (questionsToInsert != null && questionsToInsert.size() > 0) {
+            questionRepo.saveAll(questionsToInsert);
+            testQuestionRepo.saveAll(testQuestions(questionsToInsert, testId));
+        }
+
+        if (answersToInsert != null && answersToInsert.size() > 0) {
+            choiceAnswerRepo.saveAll(answersToInsert);
+        }
+
+        if (questionsToUpdate != null && questionsToUpdate.size() > 0) {
+            questionRepo.updateAll(questionsToUpdate);
+        }
+
+        if (answersToUpdate != null && answersToUpdate.size() > 0) {
+            choiceAnswerRepo.updateAll(answersToUpdate);    
+        }
         
         return resFactory.success(TestResMsg.TEST_UPDATE_SUCCESS, null);
     }
